:sparkles: feat(): Adicionando integração para criar comandos
