+++ client/src/pages/ListDevices/ListDevices.tsx
@@ -2,9 +2,6 @@
 import { UseMobile } from "@contexts/MobileContext";
 import {
   Button,
-  IconButton,
-  Modal,
-  OutlinedInput,
   Pagination,
   Stack,
   Table,
@@ -20,14 +17,10 @@ import { useEffect, useState } from "react";
 import TableCell, { tableCellClasses } from "@mui/material/TableCell";
 import { styled } from "@mui/material/styles";
 
-// Icons
-import ClearIcon from "@mui/icons-material/Clear";
-import DeleteIcon from "@mui/icons-material/Delete";
-import AddIcon from "@mui/icons-material/Add";
-import { Command } from "@models/Command";
 import { ModalNewDevice } from "@pages/NewDevice/NewDevice";
 import { DeviceService } from "@services/deviceService";
 import { Device } from "@models/Device";
+import { ModalNewCommands } from "@pages/CreateCommand/CreateCommand";
 
 const StyledTableCell = styled(TableCell)(() => ({
   [`&.${tableCellClasses.head}`]: {
@@ -56,39 +49,9 @@ export default function ListDevices() {
   const [devices, setDevices] = useState<Device[]>([])
   const [page, setPage] = useState(1);
   const [rowsPerPage] = useState(6);
-  const [selectedDevice, setSelectedDevice] = useState<number | null>(null);
+  const [selectedDevice, setSelectedDevice] = useState<number>(0);
   const [openModal, setOpenModal] = useState(false);
   const [openModalNewDevice, setOpenModalNewDevice] = useState(false);
-
-  const [formDataCommands, setFormDataCommands] = useState<Command[]>([]);
-
-  const handleAddCommand = () => {
-    const newCommand = {
-      id: 0,
-      name: "",
-      url: "",
-    };
-    setFormDataCommands([...formDataCommands, newCommand]);
-  };
-
-  const handleUpdateCommand = (
-    index: number,
-    field: keyof Command,
-    value: any
-  ) => {
-    const updatedCommands = [...formDataCommands];
-    updatedCommands[index] = {
-      ...updatedCommands[index],
-      [field]: value,
-    };
-    setFormDataCommands(updatedCommands);
-  };
-
-  const handleRemoveCommand = (index: number) => {
-    const updatedCommands = [...formDataCommands];
-    updatedCommands.splice(index, 1);
-    setFormDataCommands(updatedCommands);
-  };
 
   const handleChangePage = (
     _event: React.ChangeEvent<unknown>,
@@ -104,7 +67,7 @@ export default function ListDevices() {
 
   const handleCloseModal = () => {
     setOpenModal(false);
-    setSelectedDevice(null);
+    setSelectedDevice(0);
   };
 
   const handleOpenModalNewDevice = () => {
@@ -254,173 +217,16 @@ export default function ListDevices() {
           size="small"
         />
       </Stack>
-      <Modal open={openModal} onClose={handleCloseModal}>
-        <Stack
-          sx={{
-            position: "absolute",
-            top: "50%",
-            left: "50%",
-            transform: "translate(-50%, -50%)",
-            width: isMobile ? "100%" : "50%",
-            bgcolor: "#fff",
-            p: isMobile ? "0.5rem" : "2rem",
-            borderRadius: "8px",
-            height: isMobile ? "100%" : "35rem",
-            paddingTop: isMobile ? "2rem" : "2rem",
-            border: "none",
-            overflowY: "scroll",
-          }}
-        >
-          <Stack
-            sx={{
-              width: "100%",
-              flexDirection: "row",
-              justifyContent: "space-between",
-            }}
-          >
-            <Typography
-              sx={{
-                fontSize: "1.2rem",
-                fontWeight: 600,
-              }}
-            >
-              Dispositivo: {selectedDevice}
-              <Typography
-                sx={{
-                  fontSize: "0.8rem",
-                  fontWeight: 400,
-                }}
-              >
-                Ao salvar você adicionará esse dispositivo a sua lista pessoal
-              </Typography>
-            </Typography>
-            <IconButton onClick={handleCloseModal}>
-              <ClearIcon />
-            </IconButton>
-          </Stack>
-          <Stack marginTop="3rem" height="auto">
-            <Stack>
-              <Typography
-                sx={{
-                  fontSize: "1.2rem",
-                  fontWeight: 500,
-                }}
-              >
-                Adicionar comandos
-              </Typography>
-            </Stack>
-            <Stack>
-              {formDataCommands.map((command, index) => (
-                <>
-                  <Stack mt="2rem" width="100%">
-                    <Typography>Comando nº {index + 1}</Typography>
-                  </Stack>
-                  <Stack
-                    key={index}
-                    width="100%"
-                    flexDirection={isMobile ? "column" : "row"}
-                    mt="0.5rem"
-                    sx={{
-                      display: "flex",
-                      alignItems: "center",
-                      justifyContent: "space-around",
-                    }}
-                  >
-                    <Stack width={isMobile ? "100%" : "45%"}>
-                      <OutlinedInput
-                        placeholder="Digite o nome do comando"
-                        value={command.name}
-                        onChange={(e) =>
-                          handleUpdateCommand(index, "name", e.target.value)
-                        }
-                      />
-                    </Stack>
-                    <Stack
-                      width={isMobile ? "100%" : "45%"}
-                      mt={isMobile ? "1.5rem" : 0}
-                    >
-                      <OutlinedInput
-                        placeholder="Digite a URL"
-                        value={command.url}
-                        onChange={(e) =>
-                          handleUpdateCommand(index, "url", e.target.value)
-                        }
-                      />
-                    </Stack>
-                    <Stack
-                      width={isMobile ? "100%" : "5%"}
-                      mt={isMobile ? "1rem" : 0}
-                      alignItems={isMobile ? "end" : "center"}
-                      justifyContent={isMobile ? "end" : "center"}
-                      flexDirection={"row"}
-                    >
-                      <IconButton onClick={() => handleRemoveCommand(index)}>
-                        <DeleteIcon />
-                      </IconButton>
-                    </Stack>
-                  </Stack>
-                </>
-              ))}
-            </Stack>
-            <Stack width="100%" alignItems="end" mt="2rem">
-              <Button
-                onClick={handleAddCommand}
-                sx={{
-                  width: "15rem",
-                  background: "#808080	",
-                  color: "#FFFFFF",
-                  "&:hover": {
-                    background: "#FFFFFF",
-                    color: "#808080	",
-                    border: "1px solid #556B2F",
-                  },
-                }}
-              >
-                <AddIcon fontSize="medium" /> Novo comando
-              </Button>
-            </Stack>
-          </Stack>
-          <Stack
-            flexDirection={isMobile ? "column" : "row"}
-            justifyContent={isMobile ? "center" : "space-between"}
-            alignItems={"center"}
-            mt={"2rem"}
-          >
-            <Button
-              sx={{
-                width: isMobile ? "80%" : "13rem",
-                background: "#8B0000",
-                color: "#FFFFFF",
-                "&:hover": {
-                  background: "#FFFFFF",
-                  color: "#8B0000",
-                  border: "1px solid #8B0000",
-                },
-              }}
-            >
-              Cancelar
-            </Button>
-            <Button
-              sx={{
-                marginTop: isMobile ? "1rem" : 0,
-                width: isMobile ? "80%" : "13rem",
-                background: "#6B8E23",
-                color: "#FFFFFF",
-                "&:hover": {
-                  background: "#FFFFFF",
-                  color: "#6B8E23",
-                  border: "1px solid #6B8E23",
-                },
-              }}
-            >
-              Salvar
-            </Button>
-          </Stack>
-        </Stack>
-      </Modal>
+
       <ModalNewDevice
         handleCloseModalNewDevice={handleCloseModalNewDevice}
         openModalNewDevice={openModalNewDevice}
+      />
+
+      <ModalNewCommands
+        handleCloseModal={handleCloseModal}
+        openModal={openModal}
+        selectedDevice={selectedDevice}
       />
     </Stack>
   );
