+++ client/src/services/deviceService.ts
@@ -6,7 +6,7 @@ import axios, { AxiosResponse } from "axios";
 export const DeviceService = {
   Post: async (device: Device) => {
     try {
-      const API_URL = `${baseUrlApi.baseUrlApi}/device/createDevice`;
+      const API_URL = `${baseUrlApi.baseUrlApi}/devices/createDevice`;
       const httpHeaders = header();
       const response: AxiosResponse = await axios.post(API_URL, device, {
         headers: httpHeaders,
@@ -25,9 +25,9 @@ export const DeviceService = {
       }
     }
   },
-  AddCommand: async (deviceId: string, device: Device) => {
+  AddCommand: async (deviceId: number, device: Device) => {
     try {
-      const API_URL = `${baseUrlApi.baseUrlApi}/device/${deviceId}/addCommands`;
+      const API_URL = `${baseUrlApi.baseUrlApi}/devices/${deviceId}/addCommands`;
       const httpHeaders = header()
       const response: AxiosResponse = await axios.put(API_URL, device, {
         headers: httpHeaders,
@@ -42,7 +42,7 @@ export const DeviceService = {
   },
   GetAll: async () => {
     try {
-      const API_URL = `${baseUrlApi.baseUrlApi}/device/allDevices`;
+      const API_URL = `${baseUrlApi.baseUrlApi}/devices/allDevices`;
       const httpHeaders = header();
       const response: AxiosResponse = await axios.get(API_URL, {
         headers: httpHeaders,
@@ -57,7 +57,7 @@ export const DeviceService = {
   },
   GetByUserId: async () => {
     try {
-      const API_URL = `${baseUrlApi}/device/getDevicesByUser`;
+      const API_URL = `${baseUrlApi}/devices/getDevicesByUser`;
       const httpHeaders = {
         "Content-Type": "application/json",
       };
@@ -70,6 +70,23 @@ export const DeviceService = {
       }
     } catch (error) {
       throw new Error("Erro ao obter dispositivos");
+    }
+  },
+  GetById: async (deviceId: number) => {
+    try {
+      const API_URL = `${baseUrlApi}/devices/getById/${deviceId}`;
+      const httpHeaders = {
+        "Content-Type": "application/json",
+      };
+      const response = await axios.get(API_URL, {
+        headers: httpHeaders,
+      });
+
+      if (response) {
+        return response.data;
+      }
+    } catch (error) {
+      throw new Error("Erro ao obter dispositivo");
     }
   },
 };
